% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{cec.interactive}
\alias{cec.interactive}
\title{Interactive Cross-Entropy Clustering}
\usage{
cec.interactive(
  x,
  centers,
  type = c("covariance", "fixedr", "spherical", "diagonal", "eigenvalues", "all"),
  iter.max = 40,
  nstart = 1,
  param,
  centers.init = c("kmeans++", "random"),
  card.min = "5\%",
  keep.removed = FALSE,
  readline = TRUE
)
}
\description{
Internal function to run \code{\link{cec}} interactively.
}
