% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.params.R
\name{create.cec.params.for.models}
\alias{create.cec.params.for.models}
\title{Parameters for C}
\usage{
create.cec.params.for.models(k, n, type.arg, param.arg)
}
\description{
Internal function to prepare the clustering parameters for the 
 C function.
}
