% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse}
\alias{ellipse}
\title{Ellipse}
\usage{
ellipse(mean, cov, npoints = 250)
}
\description{
Internal function to compute points on an ellipse's perimeter 
 for a given mean vector and covariance matrix.
}
