% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.centers.R
\name{init.centers}
\alias{init.centers}
\title{Cluster Center Initialization}
\usage{
init.centers(x, k, method = c("kmeans++", "random"))
}
\description{
Internal function to automatically initialize the centers of the
 clusters before running the Cross-Entropy Clustering algorithm.
}
