% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{mixShapes}
\alias{mixShapes}
\title{Mixed Shapes Clusters}
\description{
Matrix of 2-dimensional points that form circular and elliptical 
 patterns.
}
\examples{
data(mixShapes)
plot(mixShapes, cex = 0.5, pch = 19);

}
\keyword{datasets}
