% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.covariance.R
\name{model.covariance}
\alias{model.covariance}
\title{Model Covariance}
\usage{
model.covariance(type, cov, mean, param)
}
\description{
Internal function to extract the model covariance matrix from 
 the output of the Cross-Entropy Clustering algorithm.
}
