% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.covariance.R
\name{model.mean}
\alias{model.mean}
\title{Model Mean}
\usage{
model.mean(type, center, param)
}
\description{
Internal function to extract the model mean from the output of 
 the Cross-Entropy Clustering algorithm.
}
