% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mouseset}
\alias{mouseset}
\title{Mouse}
\usage{
mouseset(
  n = 4000,
  r.head = 2,
  r.left.ear = 1.1,
  r.right.ear = 1.1,
  left.ear.dist = 2.5,
  right.ear.dist = 2.5,
  dim = 2
)
}
\description{
Internal function to generate a cluster of points uniformly 
 distributed inside a Mickey mouse shape.
}
