\name{print.cec}
\alias{print.cec}
\title{
Print CEC.
}
\description{
Presents structure of the \code{cec} results object in the form of text.
}
\usage{
\method{print}{cec}(x, ...)
}
\arguments{
  \item{x}{
The result of \code{cec} function.
}
  \item{\dots}{
Ignored.
}
}
\seealso{
\code{\link{plot.cec}}
}
\examples{
## See the examples of function cec.
}

\keyword{ ~print }
\keyword{ ~cec }
