\name{run.cec.tests}
\alias{run.cec.tests}

\title{
CEC package tests.
}
\description{

This function is used to run "unit test"-like system for testing \code{cec} package. 
The set of tests is located in \code{inst/cec_tests} directory and is consists of .R files 
defining each test case. This is also used for R CMD check.
}

\usage{
run.cec.tests()
}

\keyword{ ~tests }
\keyword{ ~unit }
\keyword{ ~cec }
