\name{threeGaussians}
\alias{threeGaussians}
\docType{data}
\title{
threeGaussians
}
\description{
Matrix of 2-dimensional points from three Gaussians with means equal (0, 0).
}
\usage{data(threeGaussians)}

\examples{
data(threeGaussians)
plot(threeGaussians, cex = 0.5, pch = 19);
}
\keyword{datasets}
