% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{Tset}
\alias{Tset}
\title{T-Shaped Clusters}
\description{
Matrix of 2-dimensional points that form the letter T.
}
\examples{
data(Tset)
plot(Tset, cex = 0.5, pch = 19);

}
\keyword{datasets}
