\name{cec-package}
\alias{cec-package}
\docType{package}
\title{
Cross-Entropy Clustering
}
\description{
The package presents clustering method called CEC based on Cross-Entropy Clustering theory. 
It introduces fast and highly applicable algorithm in the form of customizable function \code{cec}.
}
\details{
\tabular{ll}{
Package: \tab cec\cr
Type: \tab Package\cr
Version: \tab 0.9.1\cr
Date: \tab 2014-08-16\cr
License: \tab GPL-3 \cr
}
}
\author{
Konrad Kamieniecki
}
\references{
Spurek, P. and Tabor, J. (2014)
Cross-Entropy Clustering
\emph{Pattern Recognition} \bold{47, 9} 3046--3059
}
\seealso{
\code{\link{cec}.}
}
\keyword{ clustering, entropy, gaussian, kmeans }
