\name{CEC-package}
\alias{CEC-package}
\docType{package}
\title{
Cross-Entropy Clustering
}
\description{
Cross-Entropy Clustering (CEC) divides the data into Gaussian type clusters. It performs the automatic reduction of unnecessary clusters, while at the same time allows the simultaneous use of various type Gaussian mixture models. It's implemented in the form of a customizable function \code{cec}.
}
\details{
\tabular{ll}{
Package: \tab CEC\cr
Type: \tab Package\cr
Version: \tab 0.9.1\cr
Date: \tab 2014-10-06\cr
License: \tab GPL-3 \cr
}
}
\author{
Konrad Kamieniecki
}

\seealso{
\code{\link{cec}.}
}
\keyword{ clustering, entropy, gaussian, kmeans }
