\name{CEC-package}
\alias{CEC-package}
\docType{package}
\title{
Cross-Entropy Clustering
}
\description{
CEC divides data into Gaussian type clusters. The implementation allows the simultaneous use of various type Gaussian mixture models, performs the reduction of unnecessary clusters and it's able to discover new groups. Based on Spurek, P. and Tabor, J. (2014) <doi:10.1016/j.patcog.2014.03.006> \code{cec}.
}
\details{
\tabular{ll}{
Package: \tab CEC\cr
Type: \tab Package\cr
Version: \tab 0.10.2\cr
Date: \tab 2018-07-26\cr
License: \tab GPL-3 \cr
}
}
\author{
Konrad Kamieniecki
}
\seealso{
\code{\link{cec}}
}

\keyword{cluster}
\keyword{models}
\keyword{multivariate}
\keyword{package}
\concept{entropy}
\concept{gaussian}
\concept{gaussian mixture models}
\concept{kmeans}
