% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEC-package.R
\docType{package}
\name{CEC-package}
\alias{CEC}
\alias{CEC-package}
\title{Cross-Entropy Clustering}
\description{
CEC divides data into Gaussian type clusters. The implementation
 allows the simultaneous use of various type Gaussian mixture models,
 performs the reduction of unnecessary clusters and it's able to discover new
 groups. Based on Spurek, P. and Tabor, J. (2014) <doi:10.1016/j.patcog.2014.03.006>
 \code{cec}.
}
\seealso{
\code{\link{cec}}
}
\author{
Konrad Kamieniecki
}
\keyword{cluster}
\keyword{internal}
\keyword{models}
\keyword{multivariate}
\keyword{package}
