\name{print.cec}
\alias{print.cec}

\title{
Print CEC.
}

\description{
Presents a structure of the \code{cec} results object in the form of text.
}

\usage{
\method{print}{cec}(x, ...)
}

\arguments{

\item{x}{
Result of the \code{cec} function.
}

\item{\dots}{
Ignored.
}
}

\value{Returns the summary of the \code{cec} function result as text.}

\seealso{
\code{\link{plot.cec}}
}

\examples{
## See the examples of function cec.
}
