% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cec.R
\name{print.cec}
\alias{print.cec}
\title{Printing Cross Entropy Clusters}
\usage{
\method{print}{cec}(x, ...)
}
\arguments{
\item{x}{An object produced by \code{\link{cec}}.}

\item{...}{Ignored.}
}
\value{
This function returns nothing.
}
\description{
Print objects of class \code{\link{cec}}.
}
\examples{
## See the examples provided with the cec() function.

}
\seealso{
\code{\link{cec}}, \code{\link{plot.cec}}
}
\keyword{print}
