\name{run.cec.tests}
\alias{run.cec.tests}

\title{
CEC package tests.
}
\description{
This function is used to run \code{cec} package "unit test"-like system. 
The set of tests is located in \code{inst/cec_tests} directory and it consists of .R files 
defining each test case. This is also used for R CMD check.
}

\usage{
run.cec.tests()
}

\keyword{ ~tests }
\keyword{ ~unit }
\keyword{ ~cec }
