% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEC-package.R
\docType{data}
\name{threeGaussians}
\alias{threeGaussians}
\title{Three Gaussian Clusters}
\description{
Matrix of 2-dimensional points forming three Gaussian clusters.
}
\examples{
data(threeGaussians)
plot(threeGaussians, cex = 0.5, pch = 19);

}
\keyword{datasets}
